[CST-5303] refactored and added JavaDoc to CrossRef and VuFind services
