+++ dspace-api/src/main/java/org/dspace/importer/external/vufind/metadatamapping/VuFindFieldMapping.java
@@ -12,6 +12,12 @@ import javax.annotation.Resource;
 
 import org.dspace.importer.external.metadatamapping.AbstractMetadataFieldMapping;
 
+/**
+ * An implementation of {@link AbstractMetadataFieldMapping}
+ * Responsible for defining the mapping of the VuFind metadatum fields on the DSpace metadatum fields
+ * 
+ * @author Mykhaylo Boychuk (mykhaylo.boychuk@4science.com)
+ */
 @SuppressWarnings("rawtypes")
 public class VuFindFieldMapping extends AbstractMetadataFieldMapping {
 
@@ -24,6 +30,7 @@ public class VuFindFieldMapping extends AbstractMetadataFieldMapping {
      *                         the item.
      */
     @Override
+    @SuppressWarnings("unchecked")
     @Resource(name = "vufindMetadataFieldMap")
     public void setMetadataFieldMap(Map metadataFieldMap) {
         super.setMetadataFieldMap(metadataFieldMap);
