+++ dspace-api/src/main/resources/spring/spring-dspace-addon-import-services.xml
@@ -118,12 +118,16 @@
     <bean id="CrossRefImportService" class="org.dspace.importer.external.crossref.CrossRefImportMetadataSourceServiceImpl" scope="singleton">
         <property name="metadataFieldMapping" ref="CrossRefMetadataFieldMapping"/>
     </bean>
-    <bean id="CrossRefMetadataFieldMapping" class="org.dspace.importer.external.crossref.CrossRefFieldMapping"/>
+    <bean id="CrossRefMetadataFieldMapping" class="org.dspace.importer.external.crossref.CrossRefFieldMapping">
+        <property name="url" value="${crossref.url}"/>
+    </bean>
 
     <bean id="vufindImportService" class="org.dspace.importer.external.vufind.VuFindImportMetadataSourceServiceImpl" scope="singleton">
           <!-- Set to empty to use the default set of fields  -->
           <constructor-arg type="java.lang.String" value=""/>
           <property name="metadataFieldMapping" ref="vufindMetadataFieldMapping"/>
+          <property name="url" value="${vufind.url}"/>
+          <property name="urlSearch" value="${vufind.url.search}"/>
     </bean>
     <bean id="vufindMetadataFieldMapping" class="org.dspace.importer.external.vufind.metadatamapping.VuFindFieldMapping"/>
