+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/VuFindImportMetadataSourceServiceIT.java
@@ -29,8 +29,8 @@ import org.dspace.builder.CommunityBuilder;
 import org.dspace.builder.ItemBuilder;
 import org.dspace.content.Item;
 import org.dspace.importer.external.datamodel.ImportRecord;
+import org.dspace.importer.external.liveimportclient.service.LiveImportClientImpl;
 import org.dspace.importer.external.metadatamapping.MetadatumDTO;
-import org.dspace.importer.external.scopus.service.LiveImportClientImpl;
 import org.dspace.importer.external.vufind.VuFindImportMetadataSourceServiceImpl;
 import org.junit.Test;
 import org.mockito.ArgumentMatchers;
