+++ dspace/config/dspace.cfg
@@ -1638,3 +1638,4 @@ include = ${module_dir}/translator.cfg
 include = ${module_dir}/usage-statistics.cfg
 include = ${module_dir}/versioning.cfg
 include = ${module_dir}/workflow.cfg
+include = ${module_dir}/external-providers.cfg
