+++ dspace/config/modules/external-providers.cfg
@@ -1,0 +1,22 @@
+#---------------------------------------------------------------#
+#------------- EXTERNAL PROVIDER CONFIGURATIONS ----------------#
+#---------------------------------------------------------------#
+#  Configuration properties used solely by external providers   #
+#  as Scopus, Pubmed, CiNii and ect.                            #
+#---------------------------------------------------------------#
+
+
+#################################################################
+#----------------------   CrossRef   ---------------------------#
+#---------------------------------------------------------------#
+
+crossref.url = https://api.crossref.org/works
+
+#################################################################
+#----------------------   VuFind   -----------------------------#
+#---------------------------------------------------------------#
+
+vufind.url = https://vufind.org/advanced_demo/api/v1/record
+vufind.url.search = https://vufind.org/advanced_demo/api/v1/search
+
+#################################################################
